+++ .eslintrc.js
@@ -1,8 +1,0 @@
-module.exports = {
-  root: true,
-
-  extends: [
-    './index.js',
-    './typescript/index.js',
-  ],
-}