+++ scripts/constants.js
@@ -1,0 +1,9 @@
+const ON = 'error'
+const OFF = 'off'
+const WARN = 'warn'
+
+module.exports = {
+  ON,
+  OFF,
+  WARN,
+}