+++ scripts/create-conf-ts.js
@@ -1,0 +1,4 @@
+const baseConf = require('./base/conf-base-ts')
+const writeJson = require('./lib/write-json')
+
+writeJson('eslintrc-ts', baseConf)