+++ test/fixtures/test2.tsx
@@ -5,16 +5,16 @@
   The linting of this file will succeed if both blocks are merged, but must
   exist a warnig from TS, hidden by the @ts-ignore
 */
+/* eslint prettierx/options:[2,{semi:true}] */
 import * as React from 'react';
 
-type Props = { foo: typeof Foo, bar: readonly string[][] }
-type Dict = { [k: string]: React.Ref<any> }
+type Props = { foo: typeof Foo; bar: readonly string[][] };
+type Dict = { [k: string]: React.Ref<any> };
 
 class Test1 extends React.Component<Props> {
   render () {
     // no-unused-vars must be off. Comment next line to show the TS warning
     /* eslint-disable @typescript-eslint/no-var-requires */
-    // @ts-expect-error
     const React2 = require('react').createRef();
     const foo: Dict = {};
     foo.bar = React.createRef();